ISIS-2573: Xray: simplify and remove service
