+++ core/runtime/src/main/java/org/apache/isis/core/runtime/events/MetamodelEventService.java
@@ -27,8 +27,13 @@ import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Service;
 
 import org.apache.isis.applib.annotation.OrderPrecedence;
+import org.apache.isis.applib.services.confview.ConfigurationViewService;
 import org.apache.isis.applib.services.eventbus.EventBusService;
+import org.apache.isis.commons.internal.debug.xray.XrayDataModel;
+import org.apache.isis.commons.internal.debug.xray.XrayUi;
 import org.apache.isis.core.metamodel.events.MetamodelEvent;
+
+import lombok.val;
 
 /**
  * 
@@ -44,10 +49,26 @@ import org.apache.isis.core.metamodel.events.MetamodelEvent;
 public class MetamodelEventService {
     
     @Inject private EventBusService eventBusService;
-
-   // -- APP
+    @Inject private ConfigurationViewService configurationService;
 
     public void fireBeforeMetamodelLoading() {
+        
+        XrayUi.updateModel(model->{
+            
+            val root = model.getRootNode();
+            
+            val env = model.addDataNode(root, new XrayDataModel.KeyValue("Environment"));
+            configurationService.getEnvironmentProperties().forEach(item->{
+                env.getData().put(item.getKey(), item.getValue());    
+            });
+            
+            val config = model.addDataNode(root, new XrayDataModel.KeyValue("Config"));
+            configurationService.getVisibleConfigurationProperties().forEach(item->{
+                config.getData().put(item.getKey(), item.getValue());    
+            });
+            
+        });
+        
         eventBusService.post(MetamodelEvent.BEFORE_METAMODEL_LOADING);
     }
